requirements.txt
torch==2.8.0+cu126
torchvision==0.23.0+cu126
numpy==2.0.2
pillow==11.3.0